--- index.css
@@ -1,117 +1,0 @@
-.main-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    margin: 10px 20px;
-    padding-right: 20px;
-    padding-top: 20px;
-}
-
-.experiences-button-container {
-    display: grid;
-    grid-template-columns: auto auto;
-    grid-template-rows: 3;
-    gap: 20px;
-}
-
-.achievement-button {
-    grid-column: span 2;
-    padding: 30px 20px;
-}
-
-.experiences-button {
-    width: 200px;
-    padding: 70px 0 0 0;
-    display: flex;
-    justify-content: center;
-    text-align: center;
-}
-
-.experiences-button,
-.achievement-button {
-    font-size: 30px;
-    font-weight: 500;
-    background-color: #ffdba5;
-    display: flex;
-    justify-content: center;
-    border: 1px solid black;
-    border-radius: 40px;
-    text-decoration: none;
-    color: black;
-}
-
-.carousel {
-    display: flex;
-    align-content: center;
-    width: 900px;
-    overflow-x: hidden;
-    scroll-snap-type: x mandatory;
-}
-
-.carousel-item {
-    width: 900px;
-    scroll-snap-align: center;
-}
-
-.carousel-button-container {
-    position: absolute;
-    display: flex;
-    background-color: rgba(255, 255, 255, 0.7);
-    border-radius: 100px;
-    padding: 3px;
-    left: 395px;
-    top: 630px;
-}
-
-.carousel-buttons {
-    border-radius: 50%;
-    border: 0;
-    padding: 20px;
-    margin: 0 4px;
-    background-color: rgba(67, 62, 62, 0.85)
-}
-
-@media screen and (max-width: 1400px) {
-    .main-content {
-        flex-direction: column;
-        align-items: center;
-        padding-right: 0;
-    }
-
-    .experiences-button-container {
-        margin-top: 40px;
-    }
-
-    .carousel-button-container {
-        display: none;
-    }
-}
-
-@media screen and (max-width: 940px) {
-    .carousel {
-        width: 600px;
-    }
-
-    .carousel-item {
-        width: 600px;
-    }
-}
-
-@media screen and (max-width: 640px) {
-    .carousel {
-        width: 350px;
-    }
-
-    .carousel-item {
-        width: 350px;
-    }
-
-    .experiences-button {
-        width: 150px;
-    }
-
-    .experiences-button,
-    .achievement-button {
-        font-size: 24px;
-    }
-}
