--- index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./index.css">
-    <link rel="stylesheet" href="./shared_files/shared_styling/body.css">
-    <title>Social Cue</title>
-</head>
-<navbar id="navbar">
-</navbar>
-
-<body>
-    <div class="main-content">
-        <div class="carousel">
-            <img class="carousel-item" src="./shared_files/shared_images/european-shorthair-8136065_1920.jpg" alt="carousel-img">
-            <img class="carousel-item" src="./shared_files/shared_images/european-shorthair-8136065_1920.jpg" alt="carousel-img2">
-            <img class="carousel-item" src="./shared_files/shared_images/european-shorthair-8136065_1920.jpg" alt="carousel-img3">
-            <div class="carousel-button-container">
-                <button class="carousel-buttons"></button>
-                <button class="carousel-buttons"></button>
-                <button class="carousel-buttons"></button>
-            </div>
-        </div>
-        <div class="experiences-button-container" id="button-container">
-            <a href="./achievements/index.html" class="achievement-button">Achievements</a>
-            <a href="./camera/index.html" class="experiences-button">Camera</a>
-            <a href="./stories/index.html" class="experiences-button">Stories</a>
-            <a href="./scenarios/index.html" class="experiences-button">Scenarios</a>
-            <a href="./type_of_emotion/index.html" class="experiences-button">Types of Emotions</a>
-        </div>
-    </div>
-</body>
-
-<script src="./index.js"></script>
-<script src="./shared_files/navbar.js"></script>
-
-</html>
